AC model and repository Added

+++ src/main/java/com/nahalit/nahalapimanager/model/RlCustomer.java
@@ -45,7 +45,7 @@ public class RlCustomer {
   private String designation;
   private String officeAddress;
   private String customerPictureName;
-  private Integer companyNo;
+  private Long companyNo;
   private String password;
   private String religionName;
   private String cpAddress;
